+++ pysat/utils/__init__.py
@@ -11,5 +11,6 @@ for the pysat data directory structure.
 
 from pysat.utils import coords, files, time, registry, testing
 from pysat.utils._core import set_data_dir, scale_units, load_netcdf4
-from pysat.utils._core import NetworkLock
-from pysat.utils._core import generate_instrument_list
+from pysat.utils._core import NetworkLock, generate_instrument_list
+from pysat.utils._core import available_instruments
+from pysat.utils._core import display_available_instruments